corrigindo campos de quantidade e preco, routes e tela de pedidos

+++ db/migrate/20150808152443_create_products.rb
@@ -3,7 +3,7 @@ class CreateProducts < ActiveRecord::Migration
     create_table :products do |t|
       t.string :full_name
       t.integer :unity, default: 0
-      t.integer :in_stock, default: 0
+      t.decimal :in_stock, default: 0
 
       t.timestamps
     end
