+++ db/migrate/20150809232848_create_items.rb
@@ -1,7 +1,7 @@
 class CreateItems < ActiveRecord::Migration
   def change
     create_table :items do |t|
-      t.integer :quantity
+      t.decimal :quantity
       t.decimal :unit_price, scale: 2, precision: 14
       t.references :product, index: true
       t.references :purchase, index: true
