+++ db/migrate/20150823203646_create_order_items.rb
@@ -1,8 +1,8 @@
 class CreateOrderItems < ActiveRecord::Migration
   def change
     create_table :order_items do |t|
-      t.integer :quantity
-      t.integer :unit_price, scale: 2, precision: 14
+      t.decimal :quantity
+      t.decimal :unit_price, scale: 2, precision: 14
       t.references :product, index: true
       t.references :order, index: true
